# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Healer Hospital Management System API",
    "description": "Comprehensive API for managing hospital operations including appointments, staff, departments, and patient care",
    "version": "1.0.0",
    "contact": {
      "name": "Healer Development Team",
      "email": "dev@healer.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.healer.vn/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "Shift": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "fromTime": {
            "type": "string",
            "format": "date-time"
          },
          "toTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/shifts": {
      "get": {
        "summary": "Get shifts",
        "description": "Retrieve a list of work shifts",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shifts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shift"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create shift",
        "description": "Create a new work shift",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["doctorId", "roomId", "fromTime", "toTime"],
                "properties": {
                  "doctorId": {
                    "type": "string"
                  },
                  "roomId": {
                    "type": "string"
                  },
                  "fromTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "toTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shift created successfully"
          }
        }
      }
    },
...
  },
...
}